updating the controllers

+++ src/controller/ExampleController.ts
@@ -26,7 +26,7 @@ export const deleteExample = async (req: Request, res: Response) => {
 
 export const createExample = async (req: Request, res: Response) => {
   const { name, email, password, age, username } = req.body;
-  const avatar = req.file.filename;
+  const avatar = req.file?.filename;
 
   const passwordHash = await hash(password, 8);
 
@@ -45,16 +45,25 @@ export const createExample = async (req: Request, res: Response) => {
     password: passwordHash,
     age,
     username,
-    avatar,
+    avatar: avatar ? avatar : "",
   });
 
-  return res.json({ created_user: example, message: "Successfully created!" });
+  const user = {
+    id: example.id,
+    age: example.age,
+    name: example.name,
+    email: example.email,
+    avatar: example.avatar,
+    username: example.username,
+  };
+
+  return res.json({ user, message: "Successfully created!" });
 };
 
 export const updateExample = async (req: Request, res: Response) => {
   const { id } = req.params;
   const { name, email, password, age, username } = req.body;
-  const avatar = req.file.filename;
+  const avatar = req.file?.filename;
 
   deleteImage(id);
 
@@ -66,13 +75,16 @@ export const updateExample = async (req: Request, res: Response) => {
     password: passwordHash,
     age,
     username,
-    avatar,
+    avatar: avatar ? avatar : "",
   });
 
   if (example.affected === 1) {
-    const exampleUpdated = await getRepository(Example).findOne(id);
+    const user = await getRepository(Example).findOne(id, {
+      select: ["id", "age", "name", "email", "avatar", "username"],
+    });
+
     return res.json({
-      updated_user: exampleUpdated,
+      user,
       message: "Successfully updated!",
     });
   }
@@ -83,7 +95,9 @@ export const updateExample = async (req: Request, res: Response) => {
 export const selectOneExample = async (req: Request, res: Response) => {
   const { id } = req.params;
 
-  const example = await getRepository(Example).findOne(id);
+  const example = await getRepository(Example).findOne(id, {
+    select: ["id", "age", "name", "email", "avatar", "username"],
+  });
 
   if (!example) {
     return res.json({ message: "Register not found!" });
@@ -95,6 +109,7 @@ export const selectOneExample = async (req: Request, res: Response) => {
 export const selectExample = async (req: Request, res: Response) => {
   const example = await getRepository(Example).find({
     order: { id: "DESC" },
+    select: ["id", "age", "name", "email", "avatar", "username"],
   });
 
   if (example.length === 0) {
@@ -109,11 +124,14 @@ export const authenticateExample = async (req: Request, res: Response) => {
 
   const token = await authUserService({ email, password });
 
-  return res.json({ token });
+  return res.json({ token, message: "User Authenticated!" });
 };
 
 const deleteImage = async (id: string) => {
   const { avatar } = await getRepository(Example).findOne(id);
+
+  if (!avatar) return;
+
   unlink(path.resolve(__dirname, "..", "images", avatar), (err) => {
     if (err) throw err;
   });
